arm: tegra: grouper: initial port
Initial port of the grouper board definitions updated by comparing
Kai, Cardhu and Enterprise code.  The port currently doesn't take into
account the reseting of GPIO pins for the Nakasi 3G variant.  I do not
know if this is necessary, but it appears to be also done by Cardhu.
It should, however, support powering the 3G modem and suspend/resume,
but it remains untested due to unavailable hardware.

This commit almost compiles but is untested.